Update UserProductServlet.java
Fix price get from database from int to double.

+++ AE2CGroceryStore/src/main/java/DAO/ProductDAO.java
@@ -32,7 +32,7 @@ public class ProductDAO extends dbconnect.DBContext {
             while (rs.next()) {
 
                 Category Category = new Category(rs.getInt(6), rs.getString(7));
-                Product pro = new Product(rs.getInt("ProductID"), rs.getString(2), rs.getString(3), rs.getInt("Quantity"), rs.getInt("Price"), Category);
+                Product pro = new Product(rs.getInt("ProductID"), rs.getString(2), rs.getString(3), rs.getInt("Quantity"), rs.getDouble("Price"), Category);
                 list.add(pro);
             }
 
@@ -86,7 +86,7 @@ public class ProductDAO extends dbconnect.DBContext {
                 + "where CategoryID = ?";
         
         
-        
+
     }
 
     public Product getById(int productId) {
@@ -102,7 +102,7 @@ public class ProductDAO extends dbconnect.DBContext {
 
             if (rs.next()) {
                 Category cat = new Category(rs.getInt(6), rs.getString(7));
-                return new Product(rs.getInt(1), rs.getString(2), rs.getString(3), rs.getInt(4), rs.getInt(5), cat);
+                return new Product(rs.getInt(1), rs.getString(2), rs.getString(3), rs.getInt(4), rs.getDouble(5), cat);
             } else {
                 return null;
             }
@@ -145,7 +145,7 @@ public class ProductDAO extends dbconnect.DBContext {
             ResultSet rs = execSelectQuery(query, params);
 
             while (rs.next()) {
-                Product pro = new Product(rs.getInt(1), rs.getString(2), rs.getString(3), rs.getInt("Quantity"), rs.getInt("Price"), new Category(rs.getInt(6), rs.getString(7)));
+                Product pro = new Product(rs.getInt(1), rs.getString(2), rs.getString(3), rs.getInt("Quantity"), rs.getDouble("Price"), new Category(rs.getInt(6), rs.getString(7)));
                 list.add(pro);
             }
